Fix cpplint complains

+++ src/align/add.cpp
@@ -43,7 +43,7 @@ void align_add(ChunkStack &cs, Chunk *pc, size_t &max_col)
       }
       LOG_FMT(LALADD, "%s(%d): pc orig line=%zu, pc->col=%zu max_col=%zu min_col=%zu multi:%s prev->col=%zu prev->Len()=%zu %s\n",
               __func__, __LINE__, pc->GetOrigLine(), pc->GetColumn(), max_col, min_col, (prev->Is(CT_COMMENT_MULTI)) ? "Y" : "N",
-              (prev->Is(CT_COMMENT_MULTI)) ? prev->GetOrigColEnd() : (uint32_t)prev->GetColumn(), prev->Len(), get_token_name(prev->GetType()));
+              (prev->Is(CT_COMMENT_MULTI)) ? prev->GetOrigColEnd() : prev->GetColumn(), prev->Len(), get_token_name(prev->GetType()));
    }
 
    if (cs.Empty())
